Split up services and containers to avoid unnecessary imports
