+++ ldm/dream/containers.py
@@ -3,10 +3,26 @@
 """Containers module."""
 
 from dependency_injector import containers, providers
-from server.generation.containers import GeneratorContainer
-from server.signaling.containers import SignalingContainer
-from server.storage.containers import StorageContainer
+from server.models import Signal
 from server.logging.services import LogService
+from server.storage.containers import StorageContainer
+from server.storage.services import SignalQueueService
+from server.generation.containers import GeneratorContainer
+
+# An override signal service that does nothing but queue signals
+# TODO: convert signaling to use events and just don't inject any socketio stuff by default
+class SignalServiceOverride():
+  __queue: SignalQueueService
+
+  def __init__(self, queue: SignalQueueService):
+    self.__queue = queue
+    pass
+
+  def get_signal(self) -> Signal:
+    return self.__queue.get(block=True)
+
+  def emit(self, signal: Signal):
+    self.__queue.push(signal)
 
 class Container(containers.DeclarativeContainer):
   wiring_config = containers.WiringConfiguration(packages=['server'])
@@ -24,15 +40,15 @@ class Container(containers.DeclarativeContainer):
     StorageContainer
   )
 
-  signaling_package = providers.Container(
-    SignalingContainer,
-    signal_queue_service = storage_package.signal_queue_service
+  signal_service = providers.Singleton(
+    SignalServiceOverride,
+    queue = storage_package.signal_queue_service
   )
 
   generator_package = providers.Container(
     GeneratorContainer,
     config = config,
-    signal_service = signaling_package.signal_service,
+    signal_service = signal_service,
     generation_queue_service = storage_package.generation_queue_service,
     image_storage_service = storage_package.image_storage_service,
     image_intermediates_storage_service = storage_package.image_intermediates_storage_service,
